requirements.txt
asgiref==3.2.3
Django==3.0.3
django-ckeditor==5.9.0
django-js-asset==1.2.2
django-mptt==0.11.0
django-password-reset==2.0
django-taggit==1.2.0
Markdown==3.2.1
mysqlclient==1.4.2
Pillow==7.0.0
Pygments==2.5.2
pytz==2019.3
sqlparse==0.3.1
